PGCD deux methode

+++ Tp 3/Ex4/main.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
-#include <stdlib.h>
+
 
 int main()
 {
-    int n,f=1;
+    long n,f=1;
     printf("Donner n \n");
     scanf("%d",&n);
     for(int i=2 ;i<=n ; i++)
